Added packet movement and colours
Changes have been added to allow resources to have a particular colour.
Packet movement has also been added, including change in colour when it
holds a particular resource.

Colours have been added as a new parameter to ResourceUnit as a hex
string '0xRRGGBB'

There are some small problems such as the packet sprites not terminating
exactly on the centre of the target node.

When removing a node from network, make sure to call
node.sprite.destroy(); to remove the sprite from the canvas.