Added Shop Menus
Took much longer than I expected. Click on an empty space to purchase a
server, you can also click on an empty space to close a popup.

There is a change to servers, they have a limit on the total amount of
cables you can connect to it rather than x cities, y resources, z
servers.