Pause Button and Game Changes
Modifed how Dead Packets are dealt with. A node/cable deletes a dead
packet instantly instead of sending it.
(This makes it really hard to fill up a server now though)

Changed some more paramters of game progression so it shouldn't be too
long before something happens (2 minutes to be exact).

Also added a Pause Button ＼(★^∀^★)／