New additions and fixes
Additions:
Press TAB to snap the camera to a city.
Also whenever a new city spawns, text will appear on screen (press O for
demonstration).
(￣ー￣)ｂ

Also added a new variable called "totalTime" which keeps track of the
total amount of unpaused time in ms.

Fixes:
Made it so rainbow cables now continue to flash in pause mode.